CTPA-782 Endpoints all? working except POST cases/casegroup

+++ casesvc-api/src/main/java/uk/gov/ons/ctp/response/casesvc/representation/CaseGroupDTO.java
@@ -17,5 +17,5 @@ public class CaseGroupDTO {
 
   private Integer sampleId;
   
-  private Integer uprn;
+  private Long uprn;
 }
